delete playlist audio and cleanup deleted audios, refactor-move useraudio service methods to useraudio module from audio module

+++ packages/domain_data/lib/src/entities/playlist_audio/api/playlist_audio_local_repository_impl.dart
@@ -1,6 +1,7 @@
 import 'package:common_models/common_models.dart';
 import 'package:sonify_storage/sonify_storage.dart';
 
+import '../../audio/util/delete_unused_local_audio.dart';
 import '../model/playlist_audio.dart';
 import '../util/playlist_audio_mapper.dart';
 import 'playlist_audio_local_repository.dart';
@@ -10,11 +11,13 @@ class PlaylistAudioLocalRepositoryImpl with ResultWrap implements PlaylistAudioL
     this._playlistAudioEntityDao,
     this._playlistAudioMapper,
     this._dbBatchProviderFactory,
+    this._deleteUnusedLocalAudio,
   );
 
   final PlaylistAudioEntityDao _playlistAudioEntityDao;
   final PlaylistAudioMapper _playlistAudioMapper;
   final DbBatchProviderFactory _dbBatchProviderFactory;
+  final DeleteUnusedLocalAudio _deleteUnusedLocalAudio;
 
   @override
   Future<Result<PlaylistAudio>> create(PlaylistAudio playlistAudios) {
@@ -46,8 +49,37 @@ class PlaylistAudioLocalRepositoryImpl with ResultWrap implements PlaylistAudioL
   }
 
   @override
-  Future<Result<int>> deleteByIds(List<String> ids) {
-    return wrapWithResult(() => _playlistAudioEntityDao.deleteByIds(ids));
+  Future<Result<int>> deleteByIds(List<String> ids) async {
+    final audioIdsRes = await wrapWithResult(() => _playlistAudioEntityDao.getAudioIdsByIds(ids));
+    if (audioIdsRes.isErr) {
+      return Result.err();
+    }
+
+    final deleteResult = await wrapWithResult(() => _playlistAudioEntityDao.deleteByIds(ids));
+
+    final deleteUnusedAudiosRes = await _deleteUnusedLocalAudio.deleteByIds(audioIdsRes.dataOrThrow);
+    if (deleteUnusedAudiosRes.isErr) {
+      return Result.err();
+    }
+
+    return deleteResult;
+  }
+
+  @override
+  Future<EmptyResult> deleteById(String id) async {
+    final audioIdRes = await wrapWithResult(() => _playlistAudioEntityDao.getAudioIdById(id));
+    if (audioIdRes.isErr || audioIdRes.dataOrNull == null) {
+      return EmptyResult.err();
+    }
+
+    final deleteRes = await wrapWithEmptyResult(() => _playlistAudioEntityDao.deleteById(id));
+
+    final deleteUnusedAudioRes = await _deleteUnusedLocalAudio.deleteById(audioIdRes.dataOrThrow!);
+    if (deleteUnusedAudioRes.isErr) {
+      return EmptyResult.err();
+    }
+
+    return deleteRes;
   }
 
   @override
